requirements.txt
Django==5.0.2
reportlab==4.1.0
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn
whitenoise
dj-database-url 